selinux: allow controlling the SELinux-state through defconfigs
Allows to control the SELinux-system through the arch-defconfigs.
This commit adds the possibility to force "Enforced" or "Permissive"
as SELinux-status as setting both of them as default and let the
user decide.

Another addition is the "Predend Enforced"-flag which
will always return "Enforced" as the current SELinux-state, required
for proper SafetyNet-support.